[JSC] Add JS-prefix to runtime/IteratorPrototype.h https://bugs.webkit.org/show_bug.cgi?id=278710
Reviewed by Yusuke Suzuki.

We'll add some stuffs on Iterator.prototype to implement bug 248650.
https://tc39.es/proposal-iterator-helpers/#sec-iteratorprototype

Let's rename the implementation file.

* Source/JavaScriptCore/CMakeLists.txt:
* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/Sources.txt:
* Source/JavaScriptCore/runtime/JSGlobalObject.cpp:
(JSC::JSGlobalObject::init):
* Source/JavaScriptCore/runtime/JSGlobalObject.h:
(JSC::JSGlobalObject::iteratorPrototype const):
* Source/JavaScriptCore/runtime/JSIteratorConstructor.cpp:
(JSC::JSIteratorConstructor::create):
(JSC::JSIteratorConstructor::finishCreation):
* Source/JavaScriptCore/runtime/JSIteratorConstructor.h:
* Source/JavaScriptCore/runtime/JSIteratorPrototype.cpp: Renamed from Source/JavaScriptCore/runtime/IteratorPrototype.cpp.
(JSC::JSIteratorPrototype::finishCreation):
(JSC::JSC_DEFINE_HOST_FUNCTION):
(JSC::JSC_DEFINE_CUSTOM_GETTER):
(JSC::JSC_DEFINE_CUSTOM_SETTER):
* Source/JavaScriptCore/runtime/JSIteratorPrototype.h: Renamed from Source/JavaScriptCore/runtime/IteratorPrototype.h.
* Source/JavaScriptCore/runtime/JSIteratorPrototypeInlines.h: Renamed from Source/JavaScriptCore/runtime/IteratorPrototypeInlines.h.
(JSC::JSIteratorPrototype::createStructure):
* Source/WebCore/bindings/js/JSDOMIterator.h:

Canonical link: https://commits.webkit.org/282852@main
